requirements.txt
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.24.0
huggingface-hub>=0.19.0
accelerate>=0.25.0
datasets>=2.14.0
scikit-learn>=1.3.0
PyPDF2>=3.0.0
